+++ src/eip712/paymaster_params.rs
@@ -1,0 +1,20 @@
+use ethers::{
+    types::{Address, Bytes},
+    utils::rlp::Encodable,
+};
+use serde::Serialize;
+
+#[derive(Serialize, serde::Deserialize, Clone, Debug, Default)]
+#[serde(rename_all(serialize = "camelCase", deserialize = "camelCase"))]
+pub struct PaymasterParams {
+    pub paymaster: Address,
+    pub paymaster_input: Bytes,
+}
+
+impl Encodable for PaymasterParams {
+    fn rlp_append(&self, stream: &mut ethers::utils::rlp::RlpStream) {
+        stream.begin_list(2);
+        stream.append(&self.paymaster.as_bytes());
+        stream.append(&self.paymaster_input.to_vec());
+    }
+}
